4: Move Status into Service-Card folder
