+++ src/app/app.component.spec.ts
@@ -3,7 +3,7 @@ import { TestBed, async } from '@angular/core/testing';
 import { AppComponent } from './app.component';
 import { MaterialModule } from './mat.module';
 import { ServiceCardComponent } from './components/service-card/service-card.component';
-import { StatusComponent } from './components/status/status.component';
+import { StatusComponent } from './components/service-card/status/status.component';
 
 describe('AppComponent', () => {
   beforeEach(async(() => {
