+++ src/app/app.module.ts
@@ -3,7 +3,7 @@ import { BrowserModule } from '@angular/platform-browser'
 import { MaterialModule } from './mat.module';
 import { NgModule } from '@angular/core'
 import { ServiceCardComponent } from './components/service-card/service-card.component'
-import { StatusComponent } from './components/status/status.component'
+import { StatusComponent } from './components/service-card/status/status.component'
 
 @NgModule({
   declarations: [AppComponent, ServiceCardComponent, StatusComponent],
